updates, mainly in Bridge
