+++ Bridge/helper/autoload.php
@@ -1,0 +1,10 @@
+<?php
+
+// Declare
+declare(strict_types=1);
+
+foreach (glob(__DIR__ . '/*.php') as $filename) {
+    if (basename($filename) != 'autoload.php') {
+        include_once $filename;
+    }
+}
