+++ Bridge/helper/definitions.php
@@ -1,0 +1,13 @@
+<?php
+
+// Declare
+declare(strict_types=1);
+
+// Definitions
+if (!defined('SMARTLOCK_MODULE_GUID')) {
+    define('SMARTLOCK_MODULE_GUID', '{37C54A7E-53E0-4BE9-BE26-FB8C2C6A3D14}');
+}
+
+if (!defined('SERVER_SOCKET_GUID')) {
+    define('SERVER_SOCKET_GUID', '{8062CF2B-600E-41D6-AD4B-1BA66C32D6ED}');
+}
