+++ Bridge/locale.json
@@ -1,0 +1,62 @@
+{
+  "translations": {
+    "de": {
+      "NUKI Bridge - A project of Ulrich Bittner": "NUKI Bridge - Ein Projekt von Ulrich Bittner",
+      "Description:": "Beschreibung:",
+      "This instance integrates the NUKI Bridge into IP-Symcon.": "Diese Instanz integriert die NUKI Bridge in IP-Symcon.",
+      "You can open and close your door lock in conjunction with the NUKI SmartLock.": "In Verbindung mit dem NUKI SmartLock können Sie Ihr Türschloss öffnen und schließen.",
+      "For further information, please refer to the documentation or the notes here in the configuration.": "Weitere Informationen entnehmen Sie bitte der Dokumentation oder den Hinweisen hier in der Konfiguration.",
+      "Configuration:": "Konfiguration:",
+      "(0) Instance information": "(0) Instanzinformationen",
+      "Symcon version:\t5.1>": "Symcon Version:\t5.1>",
+      "Module version:\t1.04-1004": "Modul Version:\t1.04-1004",
+      "Date:\t\t\tApril 21, 2019": "Datum:\t\t\t21. April 2019",
+      "Time:\t\t\t10:00": "Uhrzeit:\t\t\t10:00",
+      "Developer:\t\tUlrich Bittner": "Entwickler:\t\tUlrich Bittner",
+      "Prefix:\t\t\tNUKI": "Präfix:\t\t\tNUKI",
+      "Please enter the IP-Address of your bridge:": "Bitte geben Sie die IP-Adresse Ihrer Bridge an:",
+      "IP-Address": "IP-Adresse",
+      "Please enter the Port of your bridge:": "Bitte geben Sie den Port Ihrer Bridge an:",
+      "Please enter your API Token:": "Bitte geben Sie Ihren API Token an:",
+      "Please select the category for your SmartLocks:": "Bitte wählen Sie die Kategorie für die SmartLocks aus:",
+      "Category": "Kategorie",
+      "With a callback you receive information from your bridge.": "Mit einem Callback empfangen Sie Informationen von Ihrer Bridge.",
+      "Please select if you want to use a callback:": "Bitte wählen Sie aus, ob Sie einen Callback nutzen wollen:",
+      "Use Callback": "Callback benutzen",
+      "Please enter the IP-Address of your Symcon server:": "Bitte geben Sie die IP-Adresse Ihres Symcon Servers an:",
+      "Please enter a free Port of your Symcon server:": "Bitte geben Sie einen freien Port Ihres IP-Symcon Servers an:",
+      "Please specify the callback ID for changes or deletions:": "Bitte geben Sie die Callback ID für Änderungen oder Löschungen an:",
+      "After accepting possible changes, you can use the ADD button under point (3) Callback Socket to create a callback": "Nach Übernahme eventueller Änderungen können Sie die Schaltfläche ANLEGEN unter Punkt (3) Callback Socket nutzen, um einen Callback anzulegen.",
+      "With the button LIST under point (3) Callback Socket, you can display the created callbacks.": "Mit der Schaltfläche ANZEIGEN unter Punkt (3) Callback Socket, können Sie sich die angelegten Callbacks anzeigen lassen.",
+      "Use the DELETE button under point (3) Callback Socket to delete the callback with the specified ID.": "Mit der Schaltfläche LÖSCHEN unter Punkt (3) Callback Socket löschen Sie den Callback mit der angegebenen ID.",
+      "Discover": "Suchen",
+      "Show Info": "Info anzeigen",
+      "Show Log": "Logdatei anzeigen",
+      "Delete Log": "Logdatei löschen",
+      "Reboot": "Neustart",
+      "Factory Reset": "Werkseinstellungen",
+      "Show": "Anzeigen",
+      "Sync": "Abgleichen",
+      "Add": "Anlegen",
+      "List": "Anzeigen",
+      "Delete": "Löschen",
+      "uncalibrated": "unkalibriert",
+      "locked": "verriegelt",
+      "unlocking": "aufschließend",
+      "unlocked": "entriegelt",
+      "locking": "verschließend",
+      "unlatched": "entriegelt",
+      "unlocked (lock ‘n’ go)": "entriegelt (lock ´n´ go)",
+      "unlatching": "entsperrend",
+      "NUKI Bridge will be created.": "NUKI Bridge wird erstellt",
+      "NUKI Bridge is active.": "NUKI Bridge ist aktiv",
+      "NUKI Bridge will be deleted.": "NUKI Bridge wird gelöscht",
+      "Please edit all configuration parameters of the NUKI Bridge.": "Bitte alle Konfigurationsparameter der NUKI Bridge eintragen.",
+      "NUKI Bridge is inactive.": "NUKI Bridge ist inaktiv.",
+      "NUKI Bridge configuration has an error!": "NUKI Bridge Konfiguration hat einen Fehler!",
+      "An unknown error has occurred!": "Es ist ein unbekannter Fehler aufgetreten!",
+      "Please enter the IP address of the IP-Symcon server and the port of the NUKI server socket!": "Bitte die IP-Adresse des IP-Symcon Servers und den Port des NUKI Server Sockets eintragen!",
+      "Smart Locks have been matched / created!": "Smart Locks wurden abgeglichen / angelegt!"
+    }
+  }
+}
