--- NUKIBridge/locale.json
@@ -1,39 +1,0 @@
-{
-  "translations": {
-    "de": {
-      "NUKI Bridge Module - Version 1.03 (21.04.2018, 12:30)": "NUKI Bridge Modul - Version 1.03 (21.04.2018, 12:30)",
-      "(1) General Settings": "(1) Allgemeine Einstellungen",
-      "IP-Address": "IP-Adresse",
-      "(2) NUKI Smart Locks": "(2) NUKI Türschloss",
-      "Category": "Kategorie",
-      "Use Callback": "Callback benutzen",
-      "Discover": "suchen",
-      "Get API Token": "API Token anzeigen",
-      "Show Info": "Info anzeigen",
-      "Show Log": "Logdatei anzeigen",
-      "Delete Log": "Logdatei löschen",
-      "Reboot": "Neustart",
-      "Factory Reset": "Werkseinstellungen",
-      "Show Smart Locks": "Türschlösser anzeigen",
-      "Sync Smart Locks": "Türschlösser abgleichen",
-      "Add Callback": "Callback anlegen",
-      "List Callback": "Callback anzeigen",
-      "Delete Callback": "Callback löschen",
-      "uncalibrated": "unkalibriert",
-      "locked": "verriegelt",
-      "unlocking": "aufschließend",
-      "unlocked": "entriegelt",
-      "locking": "verschließend",
-      "unlatched": "entriegelt",
-      "unlocked (lock ‘n’ go)": "entriegelt (lock ´n´ go)",
-      "unlatching": "entsperrend",
-      "NUKI Bridge will be created.": "NUKI Bridge wird erstellt.",
-      "NUKI Bridge is active.": "NUKI Bridge ist aktiv.",
-      "NUKI Bridge will be deleted.": "NUKI Bridge wird gelöscht.",
-      "Please edit all configuration parameters of the NUKI Bridge.": "Bitte alle Konfigurationsparameter der NUKI Bridge eintragen.",
-      "NUKI Bridge is inactive.": "NUKI Bridge ist inaktiv.",
-      "NUKI Bridge configuration has an error!": "NUKI Bridge Konfiguration hat einen Fehler!",
-      "An unknown error has occurred!": "Es ist ein unbekannter Fehler aufgetreten!"
-    }
-  }
-}
